refactor: 북마크 기능 hook으로 분리
