+++ src/app/(marketplace)/sharing/components/SharingListSection.tsx
@@ -20,10 +20,7 @@ import {
 import { useEffect, useState } from 'react';
 import { timeFormatter } from '@/utils/timeFormetter';
 import { NonDividingList } from './NonDividingList';
-import { mypageKeys } from '@/constants/queryKey';
-import { postBookmarkedMeetingApi, deleteBookmarkedMeetingApi } from '@/apis';
-import { useToast } from '@/components/Atoms/Toast/useToast';
-import { useQueryClient } from '@tanstack/react-query';
+import { useBookmark } from '@/hooks/useBookmark';
 
 export const SharingListSection = ({
   sharingMettingList,
@@ -39,27 +36,10 @@ export const SharingListSection = ({
   }, [sharingMettingList]);
 
   const router = useRouter();
-  const { success } = useToast();
-  const queryClient = useQueryClient();
+  const { handleBookmark } = useBookmark();
+
   const onClickCard = (id: string) => {
     router.push(`/sharing/${id}`);
-  };
-
-  const handleBookmark = async (id: string, isBookmarked: boolean) => {
-    try {
-      if (isBookmarked) {
-        const response = await deleteBookmarkedMeetingApi(+id);
-        success(response.message);
-      } else {
-        const response = await postBookmarkedMeetingApi(+id);
-        success(response.message);
-      }
-      queryClient.invalidateQueries({
-        queryKey: mypageKeys.bookmarksMeeting(),
-      });
-    } catch (error) {
-      console.error('찜 처리 실패:', error);
-    }
   };
 
   if (mettingList?.length === 0) {
